+++ spec/Knp/DictionaryBundle/Dictionary/DictionaryRegistrySpec.php
@@ -3,6 +3,9 @@
 namespace spec\Knp\DictionaryBundle\Dictionary;
 
 use Knp\DictionaryBundle\Dictionary;
+use Knp\DictionaryBundle\Dictionary\CategoryDictionaryInterface;
+use Knp\DictionaryBundle\Dictionary\DictionaryRegistry;
+use Knp\DictionaryBundle\Dictionary\SimpleCategorizedDictionary;
 use PhpSpec\ObjectBehavior;
 
 class DictionaryRegistrySpec extends ObjectBehavior
@@ -84,5 +87,37 @@ class DictionaryRegistrySpec extends ObjectBehavior
             'foo' => $dictionary,
             'dictionary' => $dictionary2,
         ]);
+    }
+
+    public function it_filter_dictionaries(SimpleCategorizedDictionary $dictionary3, SimpleCategorizedDictionary $dictionary4)
+    {
+        $dictionary3->getName()->willReturn('hello');
+        $dictionary4->getName()->willReturn('world');
+        $this->set('hello', $dictionary3);
+        $this->set('world', $dictionary4);
+
+        $this->filter(function (Dictionary $dictionary, $name) {
+            return $dictionary instanceof CategoryDictionaryInterface;
+        })->shouldHaveCount(2);
+    }
+
+    public function it_filter_by_category(SimpleCategorizedDictionary $dictionary3, SimpleCategorizedDictionary $dictionary4, SimpleCategorizedDictionary $dictionary5)
+    {
+        $dictionary3->getName()->willReturn('hello');
+        $dictionary3->getCategory()->willReturn('main');
+
+        $dictionary4->getName()->willReturn('world');
+        $dictionary4->getCategory()->willReturn('main');
+
+        $dictionary5->getName()->willReturn('PHP');
+        $dictionary5->getCategory()->willReturn('external_world');
+
+        $this->set('hello', $dictionary3);
+        $this->set('world', $dictionary4);
+        $this->set('PHP', $dictionary5);
+
+        $this->filterByCategory('main')->shouldHaveCount(2);
+        $this->filterByCategory('external_world')->shouldHaveCount(1);
+        $this->filterByCategory('external_world')->shouldBeAnInstanceOf(DictionaryRegistry::class);
     }
 }