+++ spec/Knp/DictionaryBundle/Dictionary/Factory/KeyValueSpec.php
@@ -2,6 +2,7 @@
 
 namespace spec\Knp\DictionaryBundle\Dictionary\Factory;
 
+use Knp\DictionaryBundle\Dictionary\SimpleCategorizedDictionary;
 use Knp\DictionaryBundle\Dictionary\ValueTransformer;
 use PhpSpec\ObjectBehavior;
 
@@ -60,5 +61,28 @@ class KeyValueSpec extends ObjectBehavior
             'foo2' => 'bar2',
             'foo3' => 'bar3',
         ]);
+    }
+
+    public function it_creates_a_categorized_dictionary($transformer)
+    {
+        $config = [
+            'category' => 'some',
+            'content' => [
+                'foo1' => 'bar1',
+                'foo2' => 'bar2',
+                'foo3' => 'bar3',
+            ],
+        ];
+
+        $transformer->transform('bar1')->willReturn('bar1');
+        $transformer->transform('bar2')->willReturn('bar2');
+        $transformer->transform('bar3')->willReturn('bar3');
+        $transformer->transform('foo1')->willReturn('foo1');
+        $transformer->transform('foo2')->willReturn('foo2');
+        $transformer->transform('foo3')->willReturn('foo3');
+
+        $dictionary = $this->create('hello', $config);
+
+        $dictionary->shouldBeAnInstanceOf(SimpleCategorizedDictionary::class);
     }
 }